Fix when a node has no childs
When en element of type Node has no childs, the root of the node is
returned by getDirectoryAsync. This generates an infinite loop when the
same function is called. By returning the element and not the root, this
properly detects that there is no child and finishes that branch of the
tree.